move trend event query to separate file and replace date_filter func
