+++ ee/clickhouse/queries/event_query.py
@@ -3,8 +3,8 @@ from typing import Any, Dict, List, Tuple
 
 from ee.clickhouse.models.cohort import format_person_query, get_precalculated_query, is_precalculated_query
 from ee.clickhouse.models.property import filter_element, prop_filter_json_extract
-from ee.clickhouse.queries.trends.util import populate_entity_params
-from ee.clickhouse.queries.util import date_from_clause, get_time_diff, get_trunc_func_ch, parse_timestamps
+from ee.clickhouse.queries.util import parse_timestamps
+from posthog.constants import MONTHLY_ACTIVE, WEEKLY_ACTIVE
 from posthog.models import Cohort, Entity, Filter, Property, Team
 
 
@@ -20,7 +20,6 @@ class ClickhouseEventQuery(metaclass=ABCMeta):
     _should_join_distinct_ids = False
     _should_join_persons = False
     _should_round_interval = False
-    _date_filter = None
 
     def __init__(
         self,
@@ -29,7 +28,6 @@ class ClickhouseEventQuery(metaclass=ABCMeta):
         round_interval=False,
         should_join_distinct_ids=False,
         should_join_persons=False,
-        date_filter=None,
         **kwargs,
     ) -> None:
         self._filter = filter
@@ -37,7 +35,6 @@ class ClickhouseEventQuery(metaclass=ABCMeta):
         self.params = {
             "team_id": self._team_id,
         }
-        self._date_filter = date_filter
 
         self._should_join_distinct_ids = should_join_distinct_ids
         self._should_join_persons = should_join_persons
@@ -137,8 +134,6 @@ class ClickhouseEventQuery(metaclass=ABCMeta):
             return ""
 
     def _get_date_filter(self) -> Tuple[str, Dict]:
-        if self._date_filter:
-            return self._date_filter, {}
 
         parsed_date_from, parsed_date_to, date_params = parse_timestamps(filter=self._filter, team_id=self._team_id)
 
@@ -203,58 +198,3 @@ class ClickhouseEventQuery(metaclass=ABCMeta):
         )
 
         return person_id_query, cohort_filter_params
-
-
-class TrendsEventQuery(ClickhouseEventQuery):
-    _entity: Entity
-
-    def __init__(self, entity: Entity, *args, **kwargs):
-        self._entity = entity
-        super().__init__(*args, **kwargs)
-
-    def get_query(self) -> Tuple[str, Dict[str, Any]]:
-        _fields = (
-            f"{self.EVENT_TABLE_ALIAS}.timestamp as timestamp, {self.EVENT_TABLE_ALIAS}.properties as properties"
-            + (f", {self.DISTINCT_ID_TABLE_ALIAS}.person_id as person_id" if self._should_join_distinct_ids else "")
-            + (f", {self.PERSON_TABLE_ALIAS}.person_props as person_props" if self._should_join_persons else "")
-        )
-
-        date_query, date_params = self._get_date_filter()
-        self.params.update(date_params)
-
-        prop_filters = [*self._filter.properties, *self._entity.properties]
-        prop_query, prop_params = self._get_props(prop_filters)
-        self.params.update(prop_params)
-
-        entity_query, entity_params = self._get_entity_query()
-        self.params.update(entity_params)
-
-        query = f"""
-            SELECT {_fields} FROM events {self.EVENT_TABLE_ALIAS}
-            {self._get_disintct_id_query()}
-            {self._get_person_query()}
-            WHERE team_id = %(team_id)s
-            {entity_query}
-            {date_query}
-            {prop_query}
-        """
-
-        return query, self.params
-
-    def _determine_should_join_persons(self) -> None:
-        super()._determine_should_join_persons()
-        for prop in self._entity.properties:
-            if prop.type == "person":
-                self._should_join_distinct_ids = True
-                self._should_join_persons = True
-                return
-
-    def _determine_should_join_distinct_ids(self) -> None:
-        if self._entity.math == "dau":
-            self._should_join_distinct_ids = True
-            return
-
-    def _get_entity_query(self) -> Tuple[str, Dict]:
-        entity_params, entity_format_params = populate_entity_params(self._entity)
-
-        return entity_format_params["entity_query"], entity_params
