+++ ee/clickhouse/queries/trends/total_volume.py
@@ -6,7 +6,7 @@ from django.utils import timezone
 from ee.clickhouse.client import format_sql, sync_execute
 from ee.clickhouse.models.action import format_action_filter
 from ee.clickhouse.models.property import parse_prop_clauses
-from ee.clickhouse.queries.event_query import TrendsEventQuery
+from ee.clickhouse.queries.trends.trend_event_query import TrendsEventQuery
 from ee.clickhouse.queries.trends.util import (
     enumerate_time_range,
     get_active_user_params,
@@ -28,33 +28,30 @@ class ClickhouseTrendsTotalVolume:
     def _total_volume_query(self, entity: Entity, filter: Filter, team_id: int) -> Tuple[str, Dict, Callable]:
 
         interval_annotation = get_trunc_func_ch(filter.interval)
-        num_intervals, seconds_in_interval, round_interval = get_time_diff(
+        num_intervals, seconds_in_interval, _ = get_time_diff(
             filter.interval or "day", filter.date_from, filter.date_to, team_id=team_id
         )
-        _, parsed_date_to, date_params = parse_timestamps(filter=filter, team_id=team_id)
-
         aggregate_operation, join_condition, math_params = process_math(entity)
+
+        trend_event_query = TrendsEventQuery(
+            filter=filter,
+            entity=entity,
+            team_id=team_id,
+            should_join_distinct_ids=True
+            if join_condition != "" or entity.math in [WEEKLY_ACTIVE, MONTHLY_ACTIVE]
+            else False,
+        )
+        event_query, event_query_params = trend_event_query.get_query()
 
         content_sql_params = {
             "aggregate_operation": aggregate_operation,
             "timestamp": "e.timestamp",
             "interval": interval_annotation,
-            "parsed_date_from": date_from_clause(interval_annotation, round_interval),
-            "parsed_date_to": parsed_date_to,
         }
         params: Dict = {"team_id": team_id}
-        params = {**params, **math_params, **date_params}
+        params = {**params, **math_params, **event_query_params}
 
         if filter.display in TRENDS_DISPLAY_BY_VALUE:
-            event_query, event_query_params = TrendsEventQuery(
-                filter=filter,
-                entity=entity,
-                team_id=team_id,
-                date_filter="{parsed_date_from} {parsed_date_to}",
-                should_join_distinct_ids=True if join_condition != "" else False,
-            ).get_query()
-            event_query = event_query.format(**content_sql_params)
-            params = {**params, **event_query_params}
             content_sql = VOLUME_TOTAL_AGGREGATE_SQL.format(event_query=event_query, **content_sql_params)
             time_range = enumerate_time_range(filter, seconds_in_interval)
 
@@ -68,27 +65,14 @@ class ClickhouseTrendsTotalVolume:
         else:
 
             if entity.math in [WEEKLY_ACTIVE, MONTHLY_ACTIVE]:
-                event_query, event_query_params = TrendsEventQuery(
-                    filter=filter,
-                    entity=entity,
-                    team_id=team_id,
-                    date_filter="{parsed_date_from_prev_range} {parsed_date_to}",
-                    should_join_distinct_ids=True,
-                ).get_query()
-                sql_params = get_active_user_params(filter, entity, team_id)
-                params = {**params, **event_query_params}
-                event_query = event_query.format(**sql_params, parsed_date_to=parsed_date_to)
-                content_sql = ACTIVE_USER_SQL.format(event_query=event_query, **content_sql_params, **sql_params)
+                content_sql = ACTIVE_USER_SQL.format(
+                    event_query=event_query,
+                    **content_sql_params,
+                    parsed_date_to=trend_event_query.parsed_date_to,
+                    parsed_date_from=trend_event_query.parsed_date_from,
+                    **trend_event_query.active_user_params
+                )
             else:
-                event_query, event_query_params = TrendsEventQuery(
-                    filter=filter,
-                    entity=entity,
-                    team_id=team_id,
-                    date_filter="{parsed_date_from} {parsed_date_to}",
-                    should_join_distinct_ids=True if join_condition != "" else False,
-                ).get_query()
-                event_query = event_query.format(**content_sql_params)
-                params = {**params, **event_query_params}
                 content_sql = VOLUME_SQL.format(event_query=event_query, **content_sql_params)
 
             null_sql = NULL_SQL.format(
